+++ langs/rust/examples/bunnies/src/lib.rs
@@ -5,8 +5,8 @@ use wasm_experiment::*;
 
 const MAX_BUNNIES: usize = 800;
 
-// This program maps a coordinate space (16384, 12288) to the actual NDS screen
-// resolution (256, 192) in order to avoid (expensive) float operations.
+// This program maps a coordinate space (16384, 12288) to the actual Fantasy Console screen
+// resolution in order to avoid (expensive) float operations.
 const MAPPED_SCREEN_DIMENSIONS: (i16, i16) = (SCREEN_WIDTH as i16 * 64, SCREEN_HEIGHT as i16 * 64);
 const MAPPED_SCREEN_CENTER: (i16, i16) = (MAPPED_SCREEN_DIMENSIONS.0/2, MAPPED_SCREEN_DIMENSIONS.1/2);
 
@@ -27,7 +27,7 @@ impl game_state::Game for GameState {
                 MAPPED_SCREEN_CENTER.0,
                 MAPPED_SCREEN_CENTER.1,
                 ((api::rand() % 65535) - 32768) as i16 % 128, // At the aforementioned coord space,
-                ((api::rand() % 65535) - 32768) as i16 % 128  // 128 maps to the number 2
+                ((api::rand() % 65535) - 32768) as i16 % 128  // 128 maps to the number 2.0
             );
             self.num_bunnies += 1;
         }
@@ -45,7 +45,7 @@ impl game_state::Game for GameState {
                 self.bunnies[i].3 *= -1;
                 self.bunnies[i].1 += self.bunnies[i].3;
             }
-            api::sprite((self.bunnies[i].0 >> 6) as u8, (self.bunnies[i].1 >> 6) as u8, 3);
+            api::sprite((self.bunnies[i].0 >> 6) as i32, (self.bunnies[i].1 >> 6) as i32, 3);
         }
     }
 }
